requirements.txt
psutil
schedule
requests
python-dotenv
